refactor: rename to SecurityPolicy to DOMPolicy and colocate with DOM
